user service renamed. application and env params added when necessary
